+++ src/api/services/prediction.ts
@@ -5,8 +5,8 @@
 
 import { isBrowser, isDeno, isNode, isWebWorker } from "browser-or-node"
 import { GraphClient } from "../graph"
-import { CloudPrediction, EdgePrediction, Feature, FeatureValue } from "../types"
-import { featureFromValue, featureToValue } from "./feature"
+import { CloudPrediction, EdgePrediction, PlainValue, Value } from "../types"
+import { toFunctionValue, toPlainValue } from "./value"
 import { StorageService } from "./storage"
 
 export const PREDICTION_FIELDS = `
@@ -32,17 +32,16 @@ export interface CreatePredictionInput {
      */
     tag: string;
     /**
-     * Input features.
-     * This is mutually exclusive with `features`.
+     * Input values.
      */
-    inputs?: { [key: string]: FeatureValue | Feature };
+    inputs?: { [key: string]: PlainValue | Value };
     /**
-     * Do not populate convenience fields in output features.
+     * Do not populate convenience fields in output values.
      * This defaults to `false`.
      */
     rawOutputs?: boolean;
     /**
-     * Return a data URL if the output feature is smaller than this limit (in bytes).
+     * Return a data URL if an output value is smaller than this limit (in bytes).
      */
     dataUrlLimit?: number;
 }
@@ -68,7 +67,7 @@ export class PredictionService {
         const key = Math.random().toString(36).slice(2) + Math.random().toString(36).slice(2); // this doesn't have to be good
         const inputs = rawInputs ?
             await Promise.all(Object.entries(rawInputs)
-                .map(([name, value]) => featureFromValue({ storage: this.storage, value, name, minUploadSize, key })
+                .map(([name, value]) => toFunctionValue({ storage: this.storage, value, name, minUploadSize, key })
                 .then(f => ({ ...f, name })))
             ) : null;
         const { data: { prediction: { results: rawResults, ...others } } } = await this.client.query<{ prediction: CloudPrediction & EdgePrediction }>(
@@ -80,7 +79,7 @@ export class PredictionService {
             { input: { tag, inputs, client, dataUrlLimit } }
         );
         const results = rawResults && !rawOutputs ?
-            await Promise.all(rawResults.map(feature => featureToValue({ feature: feature as Feature }))) :
+            await Promise.all(rawResults.map(value => toPlainValue({ value: value as Value }))) :
             rawResults;
         const prediction = rawResults !== undefined ? { ...others, results } : others;
         return prediction;